fix: api handler 함수 수정
