+++ pages/index.tsx
@@ -1,8 +1,8 @@
 import CardBox from "@/components/CardBox";
 import Header from "@/components/Header";
+import { Card } from "@/types/types";
 import { useRouter } from "next/router";
 import React, { useEffect, useState } from "react";
-import type { Card } from "../pages/api/cards";
 
 export default function Home() {
   const router = useRouter();
